+++ spec/rails_helper.rb
@@ -46,6 +46,7 @@ RSpec.configure do |config|
   config.include CustomEmailMatchers
   config.include FormHelpers
   config.include ActivityHelpers
+  config.include ActivityFormHelpers
   config.include StripAttributes::Matchers
   config.include TimeTravelHelpers